Important commit!
Turns out Vector class discards its data when it resizes. This was causing a bug in the HDF5 writer. I re-wrote the resize() member function inside Vector class to work like the ID class, that is to keep the data when resizing.

Also some fixes in HDF5 writer. Still not ready, but at least most of the output is there and it does not block the analysis.

Almost there.....