* Changed the way the element graph determines the number of boundary nodes for each element which influences how the graph is made. Used to be hardcoded within the domain::buildEleGraph() such that specific node numbers have specific number of boundary nodes.
Now each element must define its own number of nodes.

This brings changes in the Element class (setter new setter and getter functions for the number of boundary nodes, which is otherwise set to -1), and eventually all subclasses of the element class. The domain will complain if this is not done within an element, which will bring attention to this feature.