*** This commit fixes a big bug which affects scaling of ESSI with number of elements. Right now, ESSI scales linearly with increasing number of elements (used to scale quadratically).
Problem was that HDF5 output writer assumed that extending an ID or Vector by one element was a constant-time operation, it turns out its a linear time (does not allocate like std::vector, which uses a fancy allocator to achieve performance at the expense of increased memory usage). This resulted in lots of time being wasted in generating output.

Now the main bottleneck is in the SOE solution! Which is where we want it, cause there is a lot to be done there to improve performance!

There are several widespread changes. One change includes pre-allocating resources on HDF5 output writer. For this the domain needs to know several things (number of DOFS, elements, nodes, size of output).

I also removed all mentions to deprecated HDF5Channel class, which is an old idea for implementing HDF5 output on ESSI.